refactor(team-builder): extract FormChips into lanes/identity/cells/form-chips.tsx
Pure file move from identity-lane.tsx into the new cells/ subdirectory.
Sets up the structure for the upcoming atomic-cell extractions
(SpriteSection, MetaBar, ItemCell, AbilityCell, NatureCell, TeraCell)
which all live alongside form-chips.

No behavior change — IdentityLaneReal still renders FormChips with the
same props, picker logic and FormChip-based UI are unchanged.
